+++ src/bunit.web/TestContext.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Collections.Generic;
 using Bunit.Extensions;
 using Bunit.Rendering;
 using Microsoft.AspNetCore.Components;